Consolidated app.js and server.js. Deleted prisma.js and app.js files
