--- server/config/prisma.js
@@ -1,3 +1,0 @@
-const { PrismaClient } = require("@prisma/client")
-const prisma = new PrismaClient();
-module.exports = prisma;
